Using private-route-base on reporting
