+++ packages/reporting/src/components/auth/private-route.tsx
@@ -1,11 +1,9 @@
-import Debug from 'debug';
 import React from 'react';
 import { Unauthorized } from 'react-components';
 import { Redirect, Route, RouteProps, useLocation } from 'react-router';
+import { PrivateRouteBase } from 'rmf-auth';
 import { LOGIN_ROUTE } from '../../util/url';
 import { UserContext } from '../auth-contexts';
-
-const debug = Debug('PrivateRoute');
 
 export interface PrivateRouteProps extends React.PropsWithChildren<RouteProps> {
   // if true, do not redirect to login url if not authenticated
@@ -24,21 +22,18 @@ export const PrivateRoute = ({
   const user = React.useContext(UserContext);
   const location = useLocation();
 
-  function render() {
-    if (user) {
-      return children;
-    } else {
-      if (!noRedirectToLogin) {
-        debug('accessing private route while unauthenticated');
-        debug('redirecting to login page');
-        return <Redirect to={{ pathname: LOGIN_ROUTE, state: { from: location } }} />;
-      } else {
-        return <Unauthorized />;
-      }
-    }
-  }
-
-  return <Route {...rest}>{render()}</Route>;
+  return (
+    <Route {...rest}>
+      <PrivateRouteBase
+        user={user}
+        noRedirectToLogin={noRedirectToLogin}
+        unauthorized={<Unauthorized />}
+        redirect={<Redirect to={{ pathname: LOGIN_ROUTE, state: { from: location } }} />}
+      >
+        {children}
+      </PrivateRouteBase>
+    </Route>
+  );
 };
 
 export default PrivateRoute;
