+++ packages/rmf-auth/lib/components/index.ts
@@ -1,4 +1,4 @@
 export * from './login-card';
 export * from './login-page';
 export * from './login-base';
-export * from './private-route';
+export * from './private-route-base';
