+++ packages/rmf-auth/lib/components/private-route-base.tsx
@@ -8,7 +8,6 @@ const debug = Debug('PrivateRoute');
 export interface PrivateRouteBaseProps extends React.PropsWithChildren<RouteProps> {
   // if true, do not redirect to login url if not authenticated
   noRedirectToLogin?: boolean;
-  loginRoute?: string;
   unauthorized: React.ReactElement;
   user?: User | null;
   redirect: React.ReactElement;
@@ -21,36 +20,25 @@ export interface PrivateRouteBaseProps extends React.PropsWithChildren<RouteProp
 export const PrivateRouteBase = ({
   noRedirectToLogin,
   children,
-  loginRoute,
   unauthorized,
   redirect,
   user,
-  ...rest
 }: PrivateRouteBaseProps): React.ReactElement | null => {
-  // const user = React.useContext(UserContext);
-  // const location = useLocation();
+  function render() {
+    if (user) {
+      return children;
+    } else {
+      if (!noRedirectToLogin) {
+        debug('accessing private route while unauthenticated');
+        debug('redirecting to login page');
+        return redirect;
+      } else {
+        return unauthorized;
+      }
+    }
+  }
 
-  return user
-    ? (children as React.ReactElement)
-    : !noRedirectToLogin
-    ? (redirect as React.ReactElement)
-    : (unauthorized as React.ReactElement);
-
-  // function render(): JSX.Element | null {
-  //   if (user) {
-  //     return children;
-  //   } else {
-  //     if (!noRedirectToLogin) {
-  //       debug('accessing private route while unauthenticated');
-  //       debug('redirecting to login page');
-  //       return { redirect };
-  //     } else {
-  //       return { unauthorized };
-  //     }
-  //   }
-  // }
-
-  // return render();
+  return <>{render()}</>;
 };
 
 export default PrivateRouteBase;
